Changed delete method

+++ medical-record/src/main/java/com/hoperise/medicalrecord/controller/MedicalInformationController.java
@@ -3,6 +3,8 @@ package com.hoperise.medicalrecord.controller;
 import com.hoperise.medicalrecord.model.MedicalInformation;
 import com.hoperise.medicalrecord.repository.MedicalInformationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -29,8 +31,12 @@ public class MedicalInformationController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public void delete(@PathVariable Long id) {
-        medicalInformationRepository.deleteById(id);
+    public ResponseEntity<String> delete(@PathVariable Long id) {
+        if (medicalInformationRepository.existsById(id)) {
+            medicalInformationRepository.deleteById(id);
+            return ResponseEntity.ok("Medical information deleted successfully"); // Return 200 OK if deletion is successful
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Medical information with that ID doesn't exist"); // Return 404 Not Found if entity with the specified ID does not exist
+        }
     }
-
 }
